InvoiceItem: implement "position titles" and "position subtext"
The idea is that each position can have a (bigger) "title" and some smaller
text with more details.

The XML is not changed. Instead the code will parse the value of <InvoiceItem>:
The text before "//" is treated as "position title" while text after that
marker is returned as "position subtext".

If "//" is not present in the <InvoiceItem> value, position title is empty
and the complete text will be returned as "subtext".